+++ src/lib/platforms/qzone.ts
@@ -4,6 +4,7 @@ import { useOpenShare } from '../utils';
 import h from '../utils/h-demi';
 
 export default defineComponent({
+  name: SocialPlatforms.QZONE,
   setup() {
     const handleClick = useOpenShare(SocialPlatforms.QZONE);
     return () =>