+++ src/lib/platforms/twitter.ts
@@ -4,6 +4,7 @@ import { useOpenShare } from '../utils';
 import h from '../utils/h-demi';
 
 export default defineComponent({
+  name: SocialPlatforms.TWITTER,
   setup() {
     const handleClick = useOpenShare(SocialPlatforms.TWITTER);
     return () =>