+++ src/lib/platforms/wechat.ts
@@ -4,6 +4,7 @@ import { useOpenShare } from '../utils';
 import h from '../utils/h-demi';
 
 export default defineComponent({
+  name: SocialPlatforms.WECHAT,
   setup() {
     const handleClick = useOpenShare(SocialPlatforms.WECHAT);
     return () =>