+++ CHANGELOG.md
@@ -1,7 +1,19 @@
 # Changelog
 
+## v0.4.0
+
+- Breaking change: support tree-shaking.
+
+- Breaking change: iife file now moved to `bundle/vue-share-popup.bundle.js`.
+
+- Breaking change: now the lib version build will not bundle `@popperjs/core`.
+
+## v0.3.0 - v0.3.3
+
+- Fix: Vue 2 compatibility.
+
 ## v0.2.0
 
-- Feat: support wechat share.
+- Feat: add wechat share button.
 
-- Feat: support douban share.
+- Feat: add douban share button.